+++ CHANGELOG.md
@@ -8,10 +8,13 @@ Features:
     - transpose, ctranspose extended to manage string array.
     - horzcat, vertcat extended to manage string array.
     - double extended to manage string array.
-    - tolower, toupper  extended to manage string array.
+    - tolower, toupper extended to manage string array.
+    - strtrim extended to manage string array. 
     - fileread, filewrite extended to manage string array.
     - operators ==, ~=, <, >, >=, <= extended to manage string array.
     - operator plus extended to manage string array.
+
+  * deblank builtin: removes trailing whitespace from a cell of strings, a string array or a string.
 
   * operators ==, ~=, <, >, >=, <= reworked (Compatiblity Array Sizes increased).